Refactor inGame rule decorator to just be a function [#65]
