+++ src/assistants/neutralItemDigReminder.ts
@@ -1,11 +1,11 @@
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
+import inGame from "../engine/rules/inGame";
 import Item from "../gsi-data-classes/Item";
 import PlayerItems from "../gsi-data-classes/PlayerItems";
 import Rule from "../engine/Rule";
 import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
-import RuleDecoratorInGame from "../engine/rules/RuleDecoratorInGame";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -77,9 +77,6 @@ export default [
             new Fact(lastReminderTimeTopic, time),
         ],
     }),
-].map(
-    (rule) =>
-        new RuleDecoratorInGame(
-            new RuleDecoratorConfigurable(configTopic, rule)
-        )
-);
+]
+    .map(inGame)
+    .map((rule) => new RuleDecoratorConfigurable(configTopic, rule));
