+++ src/assistants/pause.ts
@@ -1,8 +1,8 @@
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
+import inGame from "../engine/rules/inGame";
 import Rule from "../engine/Rule";
 import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
-import RuleDecoratorInGame from "../engine/rules/RuleDecoratorInGame";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -15,7 +15,7 @@ export const assistantDescription = "Plays Jeopardy music while paused";
 
 export default new RuleDecoratorConfigurable(
     configTopic,
-    new RuleDecoratorInGame(
+    inGame(
         new Rule({
             label: rules.assistant.pause,
             trigger: [topics.paused],
