+++ src/assistants/philosophersStone.ts
@@ -1,12 +1,11 @@
 /* eslint-disable max-statements */
-import { DeepReadonly } from "ts-essentials";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
+import inGame from "../engine/rules/inGame";
 import PlayerItems from "../gsi-data-classes/PlayerItems";
 import Rule from "../engine/Rule";
 import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
-import RuleDecoratorInGame from "../engine/rules/RuleDecoratorInGame";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -88,4 +87,4 @@ export default [
     }),
 ]
     .map((rule) => new RuleDecoratorConfigurable(configTopic, rule))
-    .map((rule) => new RuleDecoratorInGame(rule));
+    .map(inGame);
