+++ src/assistants/runeWisdom.ts
@@ -1,8 +1,8 @@
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
+import inGame from "../engine/rules/inGame";
 import Rule from "../engine/Rule";
 import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
-import RuleDecoratorInGame from "../engine/rules/RuleDecoratorInGame";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -15,7 +15,7 @@ export const assistantDescription = "Reminds you of wisdom rune every 7:00";
 
 const ADVANCED_WARNING = 30;
 
-export default new RuleDecoratorInGame(
+export default inGame(
     new RuleDecoratorConfigurable(
         configTopic,
         new Rule({
