+++ src/assistants/tormenter.ts
@@ -1,10 +1,10 @@
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import helper from "./assistantHelpers";
+import inGame from "../engine/rules/inGame";
 import Rule from "../engine/Rule";
 import RuleDecoratorAtMinute from "../engine/rules/RuleDecoratorAtMinute";
 import RuleDecoratorConfigurable from "../engine/rules/RuleDecoratorConfigurable";
-import RuleDecoratorInGame from "../engine/rules/RuleDecoratorInGame";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -84,5 +84,5 @@ export default [
         },
     }),
 ]
-    .map((rule) => new RuleDecoratorInGame(rule))
+    .map(inGame)
     .map((rule) => new RuleDecoratorConfigurable(configTopic, rule));
